pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["gigadot <gigadot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gcloud = "^0.18.3"
google-cloud-storage = "^2.17.0"
google-cloud-bigquery = "^3.25.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
